pyproject.toml
[project]
name = "traveller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.5",
    "httpx>=0.28.1",
    "icalendar>=6.3.1",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
